Migrate Controller View to Juju API
Starts the major refactor to use the real Juju API to get app data.

This implements adding new controllers by their host and port and pretty
finishes off the controllers tab.

Still remaining is the port of the models view and other related
utilities that haven't been ported to the new API.